WIP backend statements deletion + update
