+++ v2/src/app/components/edit-statements/edit-statements.component.html
@@ -1,0 +1,16 @@
+<div *ngIf="statements">
+  <table class="table table-hover">
+    <tr *ngFor="let statement of statements; let i = index" [attr.data-index]="i">
+      <td>
+        {{statement}}
+      </td>
+      <td>
+          <div class="btn-group">
+            <button (click)="deleteStatement(i)" class="btn btn-danger">
+              <i class="far fa-trash-alt"></i>
+            </button>
+          </div>
+      </td>
+    </tr>
+  </table>
+</div>
