--- v2/src/app/edit-statements/edit-statements.component.html
@@ -1,19 +1,0 @@
-<div *ngIf="statements">
-  <table class="table table-hover">
-    <tr *ngFor="let statement of statements; let i = index" [attr.data-index]="i" (click)=delStatement(i)>
-      <td>
-        <input type="text" class="form-control" value="{{statement}}"/>  
-      </td>
-      <td>
-          <div class="btn-group">
-            <button (click)="updateStatement(statement, i)" class="btn btn-light">
-              <i class="fas fa-pencil-alt"></i>
-            </button>
-            <button (click)="delStatement(i)" class="btn btn-danger">
-              <i class="far fa-trash-alt"></i>
-            </button>
-          </div>
-      </td>
-    </tr>
-  </table>
-</div>
