+++ v2/src/app/survey.service.ts
@@ -1,7 +1,5 @@
 import { Injectable } from '@angular/core';           // ng core
 import { HttpClient } from '@angular/common/http';    // ng<->express client
-import { Survey } from './components/index/Survey';   // Survey.ts from index
-
 
 @Injectable({
   providedIn: 'root'
@@ -14,8 +12,10 @@ export class SurveyService {
   addSurvey(survey_name, survey_kurt) {
     const obj = {
       survey_name: survey_name,
-      survey_kurt: survey_kurt
+      survey_kurt: survey_kurt,
+      statements: [ "Testing 1", "Testing 2", "Testing 3", "Testing 4"]
     };
+    console.log(obj);
     this.http.post(`${this.uri}/add`, obj)
         .subscribe(res => console.log('Done'));
   }
@@ -48,5 +48,13 @@ export class SurveyService {
     return this
               .http
               .get(`${this.uri}/delete/${id}`);
-}
+  }
+
+  deleteStatement(id, statement_id) {
+    console.log(id);
+    console.log(statement_id);
+    return this
+            .http
+            .get(`${this.uri}/delete/${id}/s/${statement_id}`);
+  }
 }
